fix: changes to pass Stage 3
Client.java - remove final from private fields to work with JCommander

Server/Main.java - add a join() so we wait for the session to finish. Ensures the AtomicBoolean passed in gets updated and we don't get stuck in an infinite loop.

Session.java - move the output.writeUTF() to before checking if the command is "exit"

DatabaseArray.java - Add a case to switch statement for the "exit" command

DatabaseArrayTest.java - removed since they're not needed anymore